+++ CHANGELOG.md
@@ -6,12 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
-## 2.1 - 2021/09/15
+## 3.0 - 2021/10/21
 
 ### Added
 
 - AI data is synchronizable between devices via the new functions `syncdata_bytes` and `synchronize`.
 - Add multi-threading support: Parallelism is always enabled for mobile targets and can optionally be enabled for web targets. An optional set of `Feature`s is introduced to pick the assets for the chosen target during the setup (see `getInputData()`). Currently the only available `Feature` is `webParallel` (if not specified, the AI will run sequentially). The asset generation and publishing has been updated accordingly. For the AI to run in parallel on the web it is required that it runs on a WebWorker.
+- Add methods to instantiate the AI from a state: `restore`.
+
+### Changed
+- The method `create` does not accept anymore a serialized state and it can only be used to create a clean AI.
 
 ## 2.0.1 - 2021/09/07